Update DemonicGorillaPlugin.java
This double checks that the gorilla is using magic or ranged when the animation for the boulder fires.

I think when they changed demonic gorillas in may, they also made them randomly use the AOE_ATTACK animation, even while attacking with melee. They also added a overhead text to them, which triggers occasionally, but doesn't have anything to do with the attack styles as far as I can tell.